requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
basicsr>=1.4.2
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0
Pillow>=9.5.0
opencv-python>=4.8.0
tqdm>=4.65.0
einops>=0.6.0
timm>=0.9.0
lpips>=0.1.4
mamba-ssm>=1.0.0
causal-conv1d>=1.0.0
